+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to
 ### Fixed
 - Make handling of usage tracking errors more robust.
   [#1787](https://github.com/OpenFn/lightning/issues/1787)
+- Fix inspector shows selected dataclip as wiped after retying workorder from
+  a non-first step
+  [#1780](https://github.com/OpenFn/lightning/issues/1780)
 
 ## [v2.0.3] 2024-02-21